Started with Entity implementation

+++ src/de/atlascore/io/protocol/play/CorePacketOutChunkData.java
@@ -22,7 +22,7 @@ public class CorePacketOutChunkData extends AbstractPacket implements PacketOutC
 	private int x, z, bitmask, tileCount;
 	private boolean fullChunk;
 	private long[] motionBlocking;
-	private int[] biomes;
+	private short[] biomes;
 	private ByteBuf data, tiles;
 	
 	public CorePacketOutChunkData() {
@@ -100,9 +100,9 @@ public class CorePacketOutChunkData extends AbstractPacket implements PacketOutC
 			} else reader.readLongArrayTag();
 		}
 		final int length = readVarInt(in);
-		biomes = new int[length];
+		biomes = new short[length];
 		for (int i = 0; i < length; i++) {
-			biomes[i] = readVarInt(in);
+			biomes[i] = (short) readVarInt(in);
 		}
 		data = in.readBytes(readVarInt(in));
 		tileCount = readVarInt(in);
